clean up frontend
