+++ mulighetsrommet-arena-adapter/src/main/resources/web/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/src/favicon.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite App</title>
+    <title>MAAM</title>
   </head>
   <body>
     <div id="root"></div>
