+++ mulighetsrommet-arena-adapter/src/main/resources/web/package.json
@@ -18,8 +18,8 @@
   "devDependencies": {
     "@types/react": "^18.0.0",
     "@types/react-dom": "^18.0.0",
-    "@vitejs/plugin-react": "^1.3.0",
+    "@vitejs/plugin-react": "^2.0.0",
     "typescript": "^4.6.3",
-    "vite": "^2.9.9"
+    "vite": "^3.0.1"
   }
 }
