--- mulighetsrommet-arena-adapter/src/main/resources/web/src/api.ts
@@ -1,4 +1,0 @@
-export const getTopics = async () =>
-  fetch("http://localhost:8084/manager/topics", {
-    method: "GET",
-  }).then((response) => response.json());
